Fix kernel-lt and ENA support in CentOS 6 base AMI
The previous packer scripts to update the CentOS base
AMIs broke CentOS 6 in two critical ways.  First, the
kernel was changed from 3.10.x (the kernel-lt packages)
to 2.6.32 (the kernel packages).  Second, I forgot to
build ENA support into the updated kernel.  This patch
addresses both issues.

First, the default kernel in /etc/sysconfig/kernel is
changed from kernel to kernel-lt, which means that any
future yum upgrades will leave the kernel-lt version as
the default in grub (someone must have been updating
grub by hand previously).  So the 3.10/2.6.32 problem
should be fixed going forward.

Second, we now build the ENA kernel module after the
reboot step, so that we make sure we build the kernel
module for the new version of the kernel.  It might
be better to build the ENA kernel module for all
installed kernel versions (ie, include the 2.6.32
module as well), but I have mixed feelings given the
limited testing and decided to skip that step for now.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>