Rescan PCI bus to allow outdated NVMe drivers to discover volumes.
With CentOS 6 AMIs that have an outdated NVMe driver, we need to force a rescan
of the PCI bus after attaching an EBS volume so the driver can attach to it. The
rescan should not be harmful for other platforms, so triggering it
unconditionally. Implementing it here as the rescan has to be done after the
boto call has been made to attach a volume (from attachVolume.py) while polling
for the device so the rescan can cause our udev rule to kick in for
/dev/disk/by-ebs-volumeid.

Signed-off-by: Raghu Raja <craghun@amazon.com>