Forcing a non-interactive upgrade with APT
When a menu.lst upgrade is available, APT presents an interactive
ncurses-style menu with options to resolve conflicts. The default
behavior when forced is to use the local menu.lst instead of pulling in
the updated file from the repository. This commit adds the necessary
debconf options to allow a non-interactive upgrade.

See https://askubuntu.com/a/147079 for reference.

Signed-off-by: Raghu Raja <craghun@amazon.com>