Refresh the APT repo metadata prior to installing build-essentials.
As part of 5f3ed72 that added the _update_packages recipe to packer's
run_list, the APT update was removed to avoid redundancy. We install
build-essentials and wget to pull down and install Chef, which in turn
runs _update_packages.  Not doing an apt-get update prior to that ends
up looking for non-existent libc and other libraries, so adding back the
redundant update.

Signed-off-by: Raghu Raja <craghun@amazon.com>