Reboot after updating packages in packer.
_update_pacakges would almost always bring in an updated kernel. Invoking the
cfncluster::default recipe without rebooting the instance right after causes
issues with recipes like the ones that install NVIDIA drivers, which pulls down
a version of kernel-devel newer than what the instance is running with at that
time.

The pause_before timeout is a bit conservative but gives plenty of time for
packer to reconnect to the instance after the reboot and before trying to run
the following provisioner.

Signed-off-by: Raghu Raja <craghun@amazon.com>