Revert "Region: Adding describe-regions api to pull regions list"
Packer support for eu-west-3 is not release yet. Hence we can't
use packer command to copy image, so for now we have to manually copy
to eu-west-3. Revert this change to make packer script
happy with region validation.

Packer has to update it's code to use describe-region api
instead of maintaining it's own list of region. Will cut issue
to packer and possibly PR.

Signed-off-by: Mohan Gandhi <mohgan@amazon.com>
This reverts commit ce50d3432556e0cd56b6a36b639f294bde975774.