avoid double execution of resources notified by or which subscribes other resources
Chef executes resources in the order they appear in a recipe. But also
provide a way for:
- resource to be notified by other resources
- resource to subscribes other resources.

The default action for a resource is "run", so it is executed
sequentially as it appears in the recipe. If this resource subscribes
another resource, it is also executed when the subscribed resource change
its state.
Same speech for the resource notified by another resource.