Refactor recipes to fix Ubuntu interface renaming in kitchen steps
The https://github.com/awslabs/cfncluster-cookbook/pull/70 added shell commands to avoid the systemd v197 interface naming
and use the ethX syntax, for the Ubuntu 16 packer template.
With this patch we are fixing the kitchen part too.

The new _default_pre recipe is shared from packer and kitchen and
includes both the _update_packages and the new _setup_network recipes.
This step will be execute before the reboot steps.

Ref:
https://www.freedesktop.org/wiki/Software/systemd/PredictableNetworkInterfaceNames/

Signed-off-by: Enrico Usai <usai@amazon.com>