Install or update AWS CLI into AMIs for all distro
When building AMIs Ubuntu 14 and 16 Linux distributions get installed
with the latest version of the AWS CLI. Amazon, CentOS 6, 7 AMIs
are built from pre-cooked AMIs with already installed a different and
older version of the AWS-CLI.
This change installs and updates the AWS CLI to tested version 1.15.40,
for all distros when building the AMIs for CfnCluster.
It also executes the AWS-CLI update during the provisioning of the
master and compute nodes.