Get MAC address from Instance Metadata avoiding to search for device name
Ubuntu 16 uses the systemd v197 interface naming so the first network
interface name is not eth0.
There was a trick in the _setup_network.rb to avoid the v197 naming
and use ethX syntax. With this commit we are also removing this part.
See https://github.com/awslabs/cfncluster-cookbook/pull/115

By using the mac metadata we are sure to get the
instance's media access control (MAC) address.
In cases where multiple network interfaces are present,
this refers to the eth0 device
(the device for which the device number is 0).

This patch solves https://github.com/awslabs/cfncluster-cookbook/issues/116

Signed-off-by: Enrico Usai <usai@amazon.com>