fix build on centos6 (packer and kitchen)
Regression was introduced with the chef update from version 12 to 14.
As described here
https://github.com/chef/chef/commit/d24976bbd8846a60d186790f38124f312e91473d
and here
https://github.com/chef/chef/commit/b2b87c35ceb2d6e05c2be2f381b3e3dbe86a3dd1
the default path_sanity configuration was changed from true to false.
This is causing an issue on CentOS 6 where the PATH variable doesn't
contains the correct locations of the devel packages needed during the
installation of the cfncluster-node dependencies.
In particular, "pip install cffi" was not able to find the ffi.h header
file which is installed a step before via "yum install libffi-devel".

Using the shell_out function, we are forcing the variable PATH to contains
all the needed locations.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>