Fix shared volume attach error when AMI has more than one NVME devices
Fix the problem occurring when the AMI has already more than one NVME
devices attached. Before the patch, when trying get the first available
block device, we weren't considering the possibility that the devices
already attached could be /dev/nvme*. Thus we were always
returning /dev/xvdb as the first available device to use to attach the
shared volume. The error arise when the AMI come with more the one
device already attached already, so that /dev/xvdb is already taken.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>