Fix issues in slurm and torque so that multiple nodes can be added with a cold start
a. Add ASG Max number of dummy-compute nodes as part of potentially available resources to both slurm and torque
b. Make stack_name available to the nodewatcher so that the daemon that controls scaledown has access to the stack's status. Scaledown process will only start once stack status is complete

Signed-off-by: Balaji Sridharan <fnubalaj@amazon.com>