Fix latest python packages dependencies in Ubuntu 14
The cloud-init package installed in Ubuntu 14 requires and old version
of urllib3 package, which is also in the dependencies chain of awscli,
boto3 and cfncluster-node.

Installing a package from the official repositories should be
preferred against installing it via pip, because the dependencies are
more likely to be preserved.

Unfortunately, in this case only the awscli is present into the
official repositories. For boto3 (and cfncluster-node too), the
installation has to be done avoiding to update urllib3 or it will
break the cloud-init packages, not allowing anymore to connect to
the remote host.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>