Allow user to force deletion of jobs on SGE
Change the default SGE configuration in order to allow regular users
to be able to force deletion of jobs

Note: Forced deletion for jobs is executed  differently depending on
whether users are Sun Grid Engine administrators or not. In case of
administrative users, the jobs are removed from the internal database
of Sun Grid Engine immediately. For regular users,  the  equivalent
of a normal qdel is executed first, and deletion is forced only if the
normal cancellation  was  unsuccessful.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>