slurm: declare nodes in separate config file and use FUTURE for dummy nodes
Slightly refactored slurm.conf in order to include a separate config files
where nodes are declared. This allows the sqswatcher to regenerate the nodes
file when it needs to add or remove an host rather than modifying based on
regexes the full slurm.conf file

Also using FUTURE state for dummy nodes so that Slurm does not try to contact
all dummy nodes at every reboot. For a cluster with max size equals 1000 this
was causing a reboot delay of about 40 seconds.

Signed-off-by: Francesco De Martino <fdm@amazon.com>