slurm.conf: set ReturnToService=1
This is useful to recover compute nodes set in DOWN state because
for some reason the master couldn't contact them when they were added

ReturnToService controls when a DOWN node will be returned to service.
A DOWN node will become available for use upon registration with a valid
configuration only if it was set DOWN due to being non-responsive. If
the node was set DOWN for any other reason (low memory, unexpected
reboot, etc.), its state will not automatically be changed. A node
registers with a valid configuration if its memory, GRES, CPU count,
etc. are equal to or greater than the values configured in slurm.conf.

Signed-off-by: Francesco De Martino <fdm@amazon.com>