Install linux-generic package in place of linux-image-extra
We were using the kernel-release as linux-image-extra package version.
In this specific moment the kernel is 3.13.0-167-generic but the
linux-image-extra-3.13.0-167-generic doesn't exist.

We do not want to install the image-extra package directly.
Instead, we are now using the linux-generic meta-package,
which will ensure that upgrades work correctly, and that supporting
packages are also installed.

Signed-off-by: Enrico Usai <usai@amazon.com>