Add recipes to install/config NICE DCV on Centos7
With this patch we are preparing the environment in the
master instance to enable the user to connect through NICE DCV
to the master instance.

NICE DCV will be installed at AMI creation time
and configured/enabled at runtime according to the user's choiches.

We added a SH script to manage the DCV session/creation for the ec2-user
and a new External Authenticator for DCV.
The external authenticator is a python WebServer, managed as a service
by supervisord, that will run in a private virtual environment.

DCV will be automatically configured to use the ext authenticator.

Some details:
+ Add utility script to renew the HTTPs certificate used by the ext authenticator
+ Add support for GPU acceleration on NICE DCV
+ Install Gnome and Xorg at AMI creation time
+ SSH configuration: Disable X11Forwarding
+ SSH configuration: Limit Ciphers and MACs
+ NICE DCV version: 2019.1-7644
+ Disable lock screen by default for all the users
  (required since default EC2 users don't have a password)

Signed-off-by: Enrico Usai <usai@amazon.com>