Increase SlurmdTimeout and UnkillableStepTimeout
UnkillableStepTimeout: The length of time, in seconds, that Slurm will wait before deciding that
processes in a job step are unkillable (after they have been signaled with SIGKILL)
The default timeout value is 60 seconds. If exceeded, the compute node will be drained to
prevent future jobs from being scheduled on the node.

SlurmdTimeouti: The interval, in seconds, that the Slurm controller waits for slurmd to respond
before configuring that node's state to DOWN.

Increased UnkillableStepTimeout to 180 seconds and SlurmdTimeout to 180 seconds to allow
scheduler to recover especially when under heavy load.

Signed-off-by: Francesco De Martino <fdm@amazon.com>