Be sure to have dcv installed at runtime
The dcv_install recipe is not called at runtime, because it is included
in the default.rb, but it is called only at AMI creation time.

At runtime the entry point is the <scheduler>_config,
that calls base_config and base_install,
so the dcv_install was not included.

From now we are calling the dcv_install at runtime case too,
the recipe is idempotent so it is not a problem.

Signed-off-by: Enrico Usai <usai@amazon.com>