Fix gpg command parsing issues
The recipe for installing the CloudWatch agent verifies that the public
key provided by Amazon has the expected fingerprint. To do this, the
output of the `gpg` command was being used. As it turns out, there are
some differences in the way the output is formatted on ubuntu18 as
opposed to the other supported distros.

On ubuntu18, the output of the command that was being used looks like
this:
```
root@ip-172-31-27-82:~# gpg --list-keys 'Amazon CloudWatch Agent'
pub   rsa2048 2017-11-14 [SC]
      937616F3450B7D806CBD9725D58167303B789C72
uid           [ unknown] Amazon CloudWatch Agent
```

On the other supported distros (namely, amazon linux in the following
example) the output looks like this:
```
[root@ip-172-31-23-223 ~]# gpg --list-keys 'Amazon CloudWatch Agent'
pub   2048R/3B789C72 2017-11-14
uid       [ unknown] Amazon CloudWatch Agent
```

Instead of trying to adjust the parsing, I thought it best to handle it
using the [gnupg](https://pypi.org/project/gnupg/) module.

*Non-technical question*: is this allowed, since we distribute the
packages we install in the cookbook via our AMIs? The licence is GNU
General Public License v3 or later (GPLv3+) (GPLv3+).

Signed-off-by: Tim Lane <tilne@amazon.com>