Fix CloudWatch agent kitchen test
This PR fixes two issues with the CloudWatch agent kitchen test:
1. The original version uses the `[[` conditional operator, which is not
provided in every shell. This was causing test failures when running on
ubuntu AMIs for that reason.
2. The original version also assumes that `jq` is installed. After
looking at the process used to select the custom AMIs used in the
kitchen tests, this is definitely not a valid assumption.

The fix uses the return code from `grep`ing the following JSON format:
```
ubuntu@ip-172-31-3-190:~$ /opt/aws/amazon-cloudwatch-agent/bin/amazon-cloudwatch-agent-ctl -a status
{
  "status": "running",
  "starttime": "2019-11-21T21:08:13+00:00",
  "version": "1.231221.0"
}
```

Signed-off-by: Tim Lane <tilne@amazon.com>