Add DCV logs to CloudWatch logs integration config
This PR implements two new pieces of functionality:
1. DCV logs are now logged via CloudWatch when enabled for a cluster
2. A script for more easily adding to and validating the JSON file used
to create the CloudWatch agent config file.

Of particular note for this review are the implications of using a
wildcard in the `file_path` value of the `log_configs` entries with
`log_stream_name`s `dcv-xsession` and `dcv-agent`. According to the
[CloudWatch agent
documentation](https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch-Agent-Configuration-File-Details.html)
this results in only data for the latest file (by modification time)
being pushed to CloudWatch. Since the wildcard used in the `file_path`s
for those two entries represent a <username>.<session ID> pair, this
motivates two questions:
1. Are simultaneous sessions supported?
2. Once a session is closed, can it be reopened?

If the answer to both of those is "no", then it's probably fine. If not,
then generating configs on the fly, regenerating a CloudWatch agent
config file, and restarting the daemon would be worth considering.

Signed-off-by: Tim Lane <tilne@amazon.com>