Fix Ganglia installation
Apache daemon is normally not running when ganglia is installed, so the
notification sent to it at the end of the installation raised an error.
Fixed by removing the notification. Additionally, the service is now
restarted rather than just started in master_base_config recipe to
cover the case when ganglia is installed with a cluster update.

Signed-off-by: ddeidda <ddeidda@amazon.com>