Move installation of AWS Linux Kernel
Move the installation of the AWS Linux Kernel after the upgrade.
The upgrade must me done before, and it is already configured to avoid APT to present an interactive ncurses-style menu with options, required by some other packages that could have been triggered if the installation of the AWS Linux kernel was done before.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>