Enable centos7 repo on alinux2 for mpich-devel
None of the amzn2 or epel yum repos available on alinux2 have
mpich-devel avaiable. This is a package we make available on all of our
other OSes, so to workaround this limitation this commit configures a
yum repo that points at the base repo avaialble in centos7, which does
have a compatible version of mpich-devel.

Signed-off-by: Tim Lane <tilne@amazon.com>