More robust downloading of SGE archive signing key
The builds for alinux2 AMIs in China are consistently failing while
building SGE. They are failing when trying to import the public key used
to sign the debian archive that is used.

I didn't use a remote_file resource because this only needs to be done
for alinux2. Doing this via a shell command allows me to keep it out of
the general-purpose recipe used for all of the other OSes.

Signed-off-by: Tim Lane <tilne@amazon.com>