Disable SSH StrictHostKeyChecking for VPC hosts
StrictHostKeyChecking configuration will be disabled only when the target host is a host inside the VPC where the cluster is created.
The value of UserKnownHostsFile to /dev/null is needed to avoid possible clashing due to the fact that, since the cluster scales up and down based on computational needs, a new host could be instantiated with the same IP of a previously terminated host.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>