Restart X server during DCV configuration
The previous version of the DCV configuration recipe was using the
`set-default` unit file command in an attempt to stop the X server. This
sets the target to boot into, but does not stop the X server. This
commit uses the isolate unit command, which stops all units not enabled
in the specified unit (including the X server).

A kitchen test is also added to ensure that local users can access the X
server. This test is recommended in the [DCV admin guide](https://docs.aws.amazon.com/dcv/latest/adminguide/setting-up-installing-linux-checks.html#checks-xserver)
when the DCV GL package has been installed.

Signed-off-by: Tim Lane <tilne@amazon.com>