Fix Intel MPI modules file handling
Intel MPI modules file for version 2019.7.217 is setting dynamically the top dir based on the modules file location.
Copying or linking the Intel module file in the default module search directory doesn't work anymore, because in this way the top dir is not setup correctly.

With this patch, the Intel modules path location is added to the modules search location and the Intel MPI modules file is renamed from "mpi" to "intelmpi", to keep backward compatibility.

New modules file location for Intel MPI
```
$ modules avail

------------------------------------------------------------------------------- /opt/intel/impi/2019.7.217/intel64/modulefiles --------------------------------------------------------------------------------
intelmpi
```

Intel MPI module loaded as before
```
$ module load intelmpi
$ module list
Currently Loaded Modulefiles:
 1) intelmpi
```

Signed-off-by: Luca Carrogu <carrogu@amazon.com>