Set custom CLI package URL for Batch kitchen tests
Enable a URL of a custom pcluster CLI to be specified via an
environment variable when running kitchen tests for the MastserServer
using Batch as the scheduler. This is useful for the release process,
where unless the version to use is overridden, the awsbatch_config
recipe will attempt to install a version of the CLI that has not been
released.

Signed-off-by: Tim Lane <tilne@amazon.com>