Fix setting for custom omnibus installation url
According to doc https://docs.chef.io/workstation/config_yml_kitchen/
`product_name` and `product_version` are the new params that work in conjunction with `download_url`, but `download_url` can only point to an executable to be installed directly (it cannot be an omnibus installation script).

For now switch back in using the deprecated parameters `require_chef_omnibus` and `chef_omnibus_url`, that is the only combination that allows to specify a custom omnibus installation url.

`chef_omnibus_root` is required since the cinc client is installed in place of chef one, and root installation folder is different for this case.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>