Retrieve master info and slurm hostname at prep_env time and save in a file
With this patch we're saving the master_ip and the master_hostname in
the configuration file of the daemons.

These values will be saved by the daemons in the DB and we're retrieving
them from the compute nodes at runtime.

We're doing a single query call at prep_env time and saving
the values in files, to be reused in subsequents calls.

Signed-off-by: Enrico Usai <usai@amazon.com>