Manually set short and fqdn hostname at instance setup
Using the hostname resource the value you pass to the hostname parameter
is used in the "hostnamectl set-hostname ..." command and the aliases
are appended after it in the /etc/hosts file.

The fqdn is the first item in the /etc/hosts file,
it means that by using the hostname resource you should use the
fqdn in the hostname parameter, but it is a problem in our integration
because the domain name contains the cluster name (that can be long)
and the "hostname" command doesn't accept more than 64 characters in some OSes.

We're now using the resource to set the short hostname
and manual actions to set the fqdn as first entry in the
/etc/hosts file and we're also notifying ohai to relad the attributes.

# Tests:
---
[ec2-user@ondemand-dynamic-t2-xlarge-1 ~]$ hostname
ondemand-dynamic-t2-xlarge-1

[ec2-user@ondemand-dynamic-t2-xlarge-1 ~]$ hostname -f
ondemand-dynamic-t2-xlarge-1.alinux2-veryveryveryveryveryverylong.pcluster

[ec2-user@ondemand-dynamic-t2-xlarge-1 ~]$ cat /etc/resolv.conf
# Generated by NetworkManager
search eu-west-1.compute.internal alinux2-veryveryveryveryveryverylong.pcluster
nameserver 10.0.0.2
options timeout:2 attempts:5

[ec2-user@ondemand-dynamic-t2-xlarge-1 ~]$ cat /etc/hosts
127.0.0.1   localhost localhost.localdomain localhost4 localhost4.localdomain4
::1         localhost6 localhost6.localdomain6
10.0.2.50 ondemand-dynamic-t2-xlarge-1.alinux2-veryveryveryveryveryverylong.pcluster ondemand-dynamic-t2-xlarge-1
---

### Ohai reload
* ohai[reload_hostname] action reload
  - re-run ohai and merge results into node attributes

# References
Hostname resource source code:
https://github.com/chef/chef/blob/master/lib/chef/resource/hostname.rb

Signed-off-by: Enrico Usai <usai@amazon.com>

Signed-off-by: Enrico Usai <usai@amazon.com>