Slurm: enable cloud_dns option when DNS is configured
Since all nodes are resolvable over DNS we can enable cloud_dns option and
leave TreeWidth to its default (50). This can improve improve communications
when jobs request lots of nodes and when there are a lot of active nodes in
the system. With TreeWidth=65533, that means that slurmctld will ping every
node individually. With TreeWidth=50, slurmctld only needs to ping 50 of them
and it will fan it out to the other nodes.

Signed-off-by: Francesco De Martino <fdm@amazon.com>