Fix condition to avoid failure when partition changes require a reboot
The patch will let the script continue also when the following error is returned by the parted command: "Error: Partition(s) Y on /dev/XXX have been written, but we have been unable to inform the kernel of the change, probably because it/they are in use. As a result, the old partition(s) will remain in use. You should reboot now before making further changes."

Signed-off-by: Luca Carrogu <carrogu@amazon.com>