Cleanup FSx Lustre defaults and helpers file
## FSx Lustre defaults Changes

* Moved FSx attributes initialization out from main platform switch case.
* Removed CentOS 6 attributes: this OS is not supported.
* Renamed ubuntu specific `repository_uri` attribute to be aligned with other OSes.
* Refactored attributes initialization by using `value_for_platform` keywork.
* Differentiated attributes according to the version.
   The attributes `version`, `kmod_url` and `client_url` are only used for CentOS 7.5 and 7.6
   `public_key` and `base_url` are used for CentOS 7.7 and Ubuntu.
* Renamed `get_rhel_kernel_minor_version` given that it's CentOS 7 specific.

## References
* Value for platform: https://docs.chef.io/dsl_recipe/#value_for_platform
* FSx client installation: https://docs.aws.amazon.com/fsx/latest/LustreGuide/install-lustre-client.html

## Improve function to retrieve Centos 7 kernel minor version

The function is in the `value_for_platform` method,
so it is called in any case.

If the os was not rhel7 the `.match` method was returning a `nul` value
so the `match(...)[1]` was failing.

With this commit we're using '7' as a default and calling the function only if needed.

In any case we're also adding robustness to the function to raise an
error if the match fails.

## Improve function to retrieve Chrony reload command

The chrony_reload_command variable could be initialized.
Now we're raising an error if the init package method is an unsupported one.

Signed-off-by: Enrico Usai <usai@amazon.com>