Stop chrony immediately; restart it eventually
This commit changes the actions taken after successfully executing the
append_if_no_line resource in chrony_install. Before the change, the
service would be reloaded immediately. After, it will be stopped
immediately, and only after the chef run ends will it be started.

The issue with the previous approach is that the service implements
`restart` and `force-reload` by stopping the daemon, sleeping for a
second, and then starting it. This breaks when, for whatever reason, the
daemon has not yet stopped when the service attempts to start it after
sleeping.

The most notable drawback of this approach is that the service won't be
running until the reload notification is delivered to the service after
the chef run ends. Given that nothing else in the cookbook depends on
this service at the moment, this seems acceptable.

Leveraging the `notifies` mechanism prevents us from having to implement
our own logic to stop the service, waiting until it's truly stopped, and
then restarting it (or erring out if a hard-coded timeout has been
surpassed).

Signed-off-by: Tim Lane <tilne@amazon.com>