Improve function to retrieve Centos 7 kernel minor version
The function is in the `value_for_platform` method,
so it is called in any case.

If the os was not rhel7 the `.match` method was returning a `nul` value
so the `match(...)[1]` was failing.

With this commit we're using '7' as a default and calling the function only if needed.

In any case we're also adding robustness to the function to raise an
error if the match fails.

Signed-off-by: Enrico Usai <usai@amazon.com>