Apply network configuration for Ubuntu18 at Chef converge time
The `reload_network_config` is a helper function that defines
for all OSes (but Ubuntu18) a `service` Chef resource that is executed at converge time.

For Ubuntu18 it was executed at compilation time because there was no a chef resource,
now by defining a `ruby_block` we're executing it a converge time like the other OSes.

Signed-off-by: Enrico Usai <usai@amazon.com>