Enable EFA driver installation on all ARM platforms but Centos8
This commit removes the `--skip-kmod` argument on ARM instances, thus enabling the EFA driver to be installed in all platforms except for Centos8, which is currently still unsupported. The pinned version of the EFA installer is under development and will be replaced as soon at the official one will be available.

Signed-off-by: ddeidda <ddeidda@amazon.com>