CentOS 8 PowerTools repo needs to be lowercase in newer kernel versions
We're enabling `PowerTools` or `powertools` according to the kernel version:
https://access.redhat.com/articles/3078#RHEL8

The name changed with 8.3 release:
https://bugzilla.redhat.com/show_bug.cgi?id=1900785

## Additional changes

I changed the `find_rhel_minor_version` to return just a number,
without the `.` character to be easier to compare it.

With this patch I'm also removing `epel-release` installation
which is already performed as a packer step.

Signed-off-by: Enrico Usai <usai@amazon.com>