Increase number of retries for kitchen instance init
Over the past few weeks there's been four kitchen test failures due to
instances not becoming ready in time. The error message given by these
failures is 'Ran out of time waiting for the server with id INSTANCE_ID
to become ready.' This commit increases the amount of time the driver
will attempt to verify that an instance is ready from 20 to 30. This
will give instances an additional 150 seconds.

Signed-off-by: Tim Lane <tilne@amazon.com>