Update EFA installer to 1.14.0
Starting from EFA 1.14.0, GDR support is enabled by default. Therefore, this commit also removes the logic to reinstall EFA if GDR is enabled.
In kitchen test, the GDR check is moved to only check official built AMIs. This does not reduce the coverage of the test, because the check had `efa_gdr_enabled?` as part of the condition to execute. `enable_efa_gdr?` checks `node['cfncluster']['enable_efa_gdr']` in dna.json. However, in .kitchen.yml a parameter with wrong name `efa_gdr_enabled` was passed in. So the check was never executed.

Signed-off-by: Hanwen <hanwenli@amazon.com>