SQUASHME: change directory service enabled to false
This commit will fix a cluster creation issue when directory service is not enabled. This bug came from the removal of `directory_service` from compute nodes' dna.json in this [PR](https://github.com/aws/aws-parallelcluster/pull/3528). With the fix, the default of `default['cluster']["directory_service"]["enabled"]` is `false`, and the directory service recipe will be skipped if `directory_service` does not present in dna.json

Signed-off-by: Hanwen <hanwenli@amazon.com>