Implement workaround for Slurm+OpenMPI when `add_node_hostnames_in_hosts_file`  is set to true
If and only if `UseEc2Hostnames` is set to `true` on CLI, both `use_private_hostname` and `add_node_hostnames_in_hosts_file` will be set to true in Cookbook. This workaround can be turned off by setting `add_node_hostnames_in_hosts_file` to `false` in `ExtraChefAttributes` on CLI.

The workaround consists in updating the /etc/hosts` file on each compute node every time a job is allocated to a node or terminates. The update is triggered by a Slurm `Prolog`/`Epilog` script that runs in the allocation phase of a job and when the job completes. A `SLURM_NODE_ALIASES` env variable is used to retrieve the list of nodes involved in the job and their private IPs.

Note that for cloudwatch_log_files.json, I have to use `use_private_hostname` as the `feature_conditions` because the python logic to handle the conditions only read from dna.json. Unfortunately, `add_node_hostnames_in_hosts_file` does not exist in the file. This compromise is not too bad because practically `add_node_hostnames_in_hosts_file` is useful only when `use_private_hostname`. So the logs should always be upload when needed.

To improve the cloudwatch_log_files.json, we could make it as a Chef template and generate the final file accordingly.

Signed-off-by: Hanwen <hanwenli@amazon.com>