[byos] Merge from develop (#1338)
* 3.1: Configure LDAP access filter in SSSD (#1312)

* 3.1: Fix configuration of LDAP access filter in SSSD (#1315)

* Disable log4j-cve-2021-44228-hotpatch service on Amazon Linux

Signed-off-by: Francesco De Martino <fdm@amazon.com>

* Update CHANGELOG.md

* Upgrade to Intel MPI Library 2021.4.0.441

Intel MPI is now part of the OneAPI package.
The standalone offline installer is available here:
https://www.intel.com/content/www/us/en/developer/articles/tool/oneapi-standalone-components.html#mpi

Relevant changes:
- The installation path doesn't contain the patch version `.441`
- The EULA file is available in the `/opt/intel/licensing` path
- The base modules folder has been renamed from `impi` to `mpi`
- Installer options have been changed
- Extracted folder can be automatically deleted through the `--remove-extracted-files` flag
- Added Qt source code file to comply with https://www.qt.io/licensing/open-source-lgpl-obligations

Signed-off-by: Enrico Usai <usai@amazon.com>

* Implement workaround for Slurm+OpenMPI when `add_node_hostnames_in_hosts_file`  is set to true

If and only if `UseEc2Hostnames` is set to `true` on CLI, both `use_private_hostname` and `add_node_hostnames_in_hosts_file` will be set to true in Cookbook. This workaround can be turned off by setting `add_node_hostnames_in_hosts_file` to `false` in `ExtraChefAttributes` on CLI.

The workaround consists in updating the /etc/hosts` file on each compute node every time a job is allocated to a node or terminates. The update is triggered by a Slurm `Prolog`/`Epilog` script that runs in the allocation phase of a job and when the job completes. A `SLURM_NODE_ALIASES` env variable is used to retrieve the list of nodes involved in the job and their private IPs.

Note that for cloudwatch_log_files.json, I have to use `use_private_hostname` as the `feature_conditions` because the python logic to handle the conditions only read from dna.json. Unfortunately, `add_node_hostnames_in_hosts_file` does not exist in the file. This compromise is not too bad because practically `add_node_hostnames_in_hosts_file` is useful only when `use_private_hostname`. So the logs should always be upload when needed.

To improve the cloudwatch_log_files.json, we could make it as a Chef template and generate the final file accordingly.

Signed-off-by: Hanwen <hanwenli@amazon.com>

* Fix the way extra.json is merged into the final dna.json (#1320)

* Remove test for access to X server when dcv-gl is not installed

Not all graphics instance support dcv-gl, hence test to ensure local users can access X server should be skipped for this case

Signed-off-by: Luca Carrogu <carrogu@amazon.com>

* Add changelog for NVIDIA and CUDA for ARM

Signed-off-by: Luca Carrogu <carrogu@amazon.com>

* Fix string used to verify Intel MPI installation on kitchen tests

```
$ module load intelmpi && mpirun --help | grep 'Version 2021.4'
Loading intelmpi version 2021.4.0
Intel(R) MPI Library, Version 2021.4  Build 20210831 (id: 758087adf)
$ echo $?
0
```

Signed-off-by: Enrico Usai <usai@amazon.com>

* Add cluster parameter disabled_on_compute_nodes and disable integration with Ad on compute nodes if it is set to true

Signed-off-by: Giacomo Marciani <mgiacomo@amazon.com>

* Improve imds access configuration

* Execute tests for user access only at AMI build time
* Skip all configurations on compute nodes where IMDS access is always enabled

Signed-off-by: Francesco De Martino <fdm@amazon.com>

* Bump version to 3.1.0 and improve changelog

Signed-off-by: Enrico Usai <usai@amazon.com>

* Do not check parallelcluster-iptables service on awsbatch clusters

Signed-off-by: Francesco De Martino <fdm@amazon.com>

* Improve changelog

Signed-off-by: Enrico Usai <usai@amazon.com>

Co-authored-by: Giacomo Marciani <mgiacomo@amazon.com>
Co-authored-by: Francesco De Martino <fdm@amazon.com>
Co-authored-by: Francesco De Martino <demartinof@icloud.com>
Co-authored-by: Enrico Usai <usai@amazon.com>
Co-authored-by: Hanwen <hanwenli@amazon.com>