Set AuthInfo=cred_expire=70
The default value is 120 seconds, reduce AuthInfo=cred_expire=70 to allow job faster requeue when nodes are not available.
The choice of 70 seconds is justified by the clustermgtd wake up time (60 secs) + 2 times an average loop time (5 seconds).
This changes with the number of nodes, anyway in the worst case scenario that the clustermgtd hasn't finished yet, the job will get rescheduled to the same compute resource and the fail over will happen at the next iteration.

Signed-off-by: chenwany <chenwany@amazon.com>