Support mounting existing FSx for Ontap and OpenZFS
Accept two more attributes to cookbook:
fsx_volume_junction_paths: It is a comma separated list. There are strings between the comma if it is FSx for Ontap and OpenZFS. There is empty if it is FSx for Lustre.
fsx_fs_types: It is a comma separated list. It tells the cookbook what mount method to use according to the file system (Lustre, Ontap, OpenZFS)

Signed-off-by: Hanwen <hanwenli@amazon.com>