Reduce the boot time DynamoDB invocation timeout from 50 minutes to 5min
The init recipe is querying Dynamo DB to retrieve the compute instance host name. If
the compute instance is unable to reach Dynamo DB, it would take 10 minutes
before timing out. It would then retry 4 more times.

This change now sets the timeouts for the aws CLI to 30 seconds for connecting
and 1 minute for reading and uses the default aws CLI max-retry limit of 2. It
also sets the shell timeout to 5 minutes.

I also removed 'hit_' prefix from helper functions as that naming convention
is no longer used.

Signed-off-by: David Pratt <davprat@amazon.com>