Generate new "Api" field to discriminate compute resources with single and multiple instance types
Note: currently run-instances compute resources don't have all the information required for fleet
so it will be possible to force usage of run-instances API for "fleet" compute resources but
not the opposite.

To be able to use "create-fleet" API for all of them we need to propagate all the information
required for create-fleet API but this requires a change in the CLI.

Cluster config with default values don't contain for example AllocationStrategy info
for queues without InstanceTypeList.

Signed-off-by: Enrico Usai <usai@amazon.com>